Moved date utilities to the same class for simplicity
